+++ scripts/supermucNG1/job_test.sh
@@ -47,8 +47,8 @@ echo 'num_nodes:' $SLURM_JOB_NUM_NODES 'ntasks:' $SLURM_NTASKS
 ulimit -Ss 2097152
 
 ORDER=${order:-4}
-module load seissol/1.3.1-oneapi25-o${ORDER}-elas-dunav-single-impi
-#module load seissol/master-oneapi25-o${ORDER}-elas-dunav-single-impi
+#module load seissol/1.3.1-oneapi25-o${ORDER}-elas-dunav-single-impi
+module load seissol/master-oneapi25-o${ORDER}-elas-dunav-single-impi
 unset KMP_AFFINITY
 
 part_file=$1
