+++ src/dynworkflow/generate_input_files_for_dr_ensemble.py
@@ -6,9 +6,9 @@
 import os
 
 import yaml
-from kinematic_models import project_fault_tractions_onto_asagi_grid
 
 from dynworkflow import generate_input_seissol_dr
+from kinematic_models import project_fault_tractions_onto_asagi_grid
 
 
 def main(args):
