+++ src/dynworkflow/infer_fault_mesh_size_and_spatial_zoom.py
@@ -8,8 +8,9 @@ import os
 import os.path
 
 import numpy as np
-from kinematic_models.multi_fault_plane import MultiFaultPlane
 from sklearn.decomposition import PCA
+
+from kinematic_models.multi_fault_plane import MultiFaultPlane
 
 
 def infer_quantities(filename, proj, mesh_size="auto"):
